Moved SystemClock_Config to main_helper.h
